@@ -32,8 +32,8 @@
         }
     ],
     "require": {
-        "php": ">=8.0",
-        "yiisoft/yii2": "^2.0",
+        "php": "^8.1",
+        "yiisoft/yii2": "^2.0.48",
         "npm-asset/leaflet": "^1.9"
     },
     "minimum-stability": "stable",
@@ -60,9 +60,6 @@
         }
     },
     "config": {
-        "platform": {
-            "php": "8.0"
-        },
         "sort-packages": true,
         "allow-plugins": {
             "yiisoft/yii2-composer": true
