@@ -1,23 +1,24 @@
 <?php
+declare(strict_types=1);
+
 /**
- * /**
  * @copyright Copyright (c) 2013-2015 2amigOS! Consulting Group LLC
- * @link http://2amigos.us
- * @license http://www.opensource.org/licenses/bsd-license.php New BSD License
+ * @link https://2amigos.us
+ * @license https://www.opensource.org/licenses/bsd-license.php New BSD License
  */
 
 namespace dosamigos\leaflet\layers;
 
-
+use yii\base\InvalidConfigException;
 use yii\web\JsExpression;
 
 /**
  * Circle a class for drawing circle overlays on a map.
  *
- * @see http://leafletjs.com/reference.html#circle
+ * @see https://leafletjs.com/reference.html#circle
  * @author Antonio Ramirez <amigo.cobos@gmail.com>
- * @link http://www.ramirezcobos.com/
- * @link http://www.2amigos.us/
+ * @link https://www.ramirezcobos.com/
+ * @link https://www.2amigos.us/
  * @package dosamigos\leaflet\layers
  */
 /**
@@ -31,24 +32,35 @@ class Circle extends Layer
     /**
      * @var float Sets the radius of a circle. Units are in meters.
      */
-    public $radius;
+    public float $radius = 0.0;
+
+    /**
+     * @throws InvalidConfigException
+     */
+    public function init(): void
+    {
+        parent::init();
+        if ($this->getLatLng() === null) {
+            throw new InvalidConfigException("'latLng' attribute cannot be empty.");
+        }
+    }
 
     /**
      * Returns the javascript ready code for the object to render
      * @return JsExpression
      */
-    public function encode()
+    public function encode(): JsExpression
     {
         $bounds = $this->getLatLng()->toArray(true);
         $radius = $this->radius;
         $options = $this->getOptions();
         $name = $this->name;
         $map = $this->map;
-        $js = $this->bindPopupContent("L.circle($bounds, $radius, $options)") . ($map !== null ? ".addTo($map);" : "");
+        $js = $this->bindPopupContent("L.circle($bounds, $radius, $options)") . ($map !== null ? ".addTo($map)" : "");
         if (!empty($name)) {
-            $js = "var $name = $js" . ($map !== null ? "" : ";");
+            $js = "var $name = $js;";
         }
+        $js .= $this->getEvents() . ($map !== null && empty($name) ? ";" : "");
         return new JsExpression($js);
     }
-
 }
