@@ -1,8 +1,10 @@
 <?php
+declare(strict_types=1);
+
 /**
  * @copyright Copyright (c) 2013-2015 2amigOS! Consulting Group LLC
- * @link http://2amigos.us
- * @license http://www.opensource.org/licenses/bsd-license.php New BSD License
+ * @link https://2amigos.us
+ * @license https://www.opensource.org/licenses/bsd-license.php New BSD License
  */
 namespace dosamigos\leaflet\layers;
 
@@ -16,7 +18,7 @@ trait NameTrait
      * will be returned as a variable. If null, then the js creation script will
      * be returned as a constructor that you can use on other object's configuration options.
      */
-    private $_name;
+    private ?string $_name = null;
 
     /**
      * Returns the name of the layer.
@@ -25,7 +27,7 @@ trait NameTrait
      *
      * @return string name of the layer.
      */
-    public function getName($autoGenerate = false)
+    public function getName(bool $autoGenerate = false): ?string
     {
         if ($autoGenerate && $this->_name === null) {
             $this->_name = LeafLet::generateName();
@@ -36,9 +38,9 @@ trait NameTrait
     /**
      * Sets the name of the layer.
      *
-     * @param string $value name of the layer.
+     * @param string|null $value name of the layer.
      */
-    public function setName($value)
+    public function setName(?string $value): void
     {
         $this->_name = $value;
     }
