@@ -1,20 +1,23 @@
 <?php
+declare(strict_types=1);
+
 /**
  * @copyright Copyright (c) 2013-2015 2amigOS! Consulting Group LLC
- * @link http://2amigos.us
- * @license http://www.opensource.org/licenses/bsd-license.php New BSD License
+ * @link https://2amigos.us
+ * @license https://www.opensource.org/licenses/bsd-license.php New BSD License
  */
 
 namespace dosamigos\leaflet\layers;
 
 use yii\base\InvalidConfigException;
+use yii\helpers\Json;
 use yii\web\JsExpression;
 
 /**
  * Popup is used to open popups in certain places of the map. For popups directly attached to an
  * object (ie [[Marker]]) better use their `popup` attribute.
  *
- * @see http://leafletjs.com/reference.html#popup
+ * @see https://leafletjs.com/reference.html#popup
  * @package dosamigos\leaflet\layers
  */
 
@@ -29,34 +32,35 @@ class Popup extends Layer
     /**
      * @var string the HTML content of the popup
      */
-    public $content;
+    public ?string $content = null;
 
     /**
      * Initializes the marker.
      * @throws \yii\base\InvalidConfigException
      */
-    public function init()
+    public function init(): void
     {
         parent::init();
-        if (empty($this->latLng)) {
+        if ($this->getLatLng() === null) {
             throw new InvalidConfigException("'latLon' attribute cannot be empty.");
         }
     }
 
     /**
      * Returns the javascript ready code for the object to render
-     * @return string
+     * @return JsExpression
      */
-    public function encode()
+    public function encode(): JsExpression
     {
         $latLon = $this->getLatLng()->encode();
         $options = $this->getOptions();
-        $name = $this->name;
+        $name = $this->getName();
         $map = $this->map;
-        $js = "L.popup($options).setLatLng($latLon).setContent('$this->content')" . ($map !== null ? ".addTo($map);" : "");
+        $js = "L.popup($options).setLatLng($latLon).setContent(" . Json::encode($this->content) . ")" . ($map !== null ? ".addTo($map)" : "");
         if (!empty($name)) {
-            $js = "var $name = $js" . ($map !== null ? "" : ";");
+            $js = "var $name = $js;";
         }
+        $js .= $this->getEvents() . ($map !== null && empty($name) ? ";" : "");
 
         return new JsExpression($js);
     }
