@@ -1,28 +1,29 @@
 <?php
+declare(strict_types=1);
+
 /**
- *
  * PopupTrait.php
  *
  * Date: 15/02/14
  * Time: 11:03
  * @author Antonio Ramirez <amigo.cobos@gmail.com>
- * @link http://www.ramirezcobos.com/
- * @link http://www.2amigos.us/
+ * @link https://www.ramirezcobos.com/
+ * @link https://www.2amigos.us/
  */
 
 namespace dosamigos\leaflet\layers;
-
 
 trait PopupTrait
 {
     /**
      * @var string the HTML content of the popup to display when clicking on the marker.
      */
-    public $popupContent;
+    public ?string $popupContent = null;
+ 
     /**
      * @var bool whether to open the popup dialog on display.
      */
-    public $openPopup = false;
+    public bool $openPopup = false;
 
     /**
      * Binds popup content (if any) to the js code to register
@@ -31,7 +32,7 @@ trait PopupTrait
      *
      * @return string
      */
-    protected function bindPopupContent($js)
+    protected function bindPopupContent(string $js): string
     {
         if (!empty($this->popupContent)) {
             $content = addslashes($this->popupContent);
