@@ -1,8 +1,10 @@
 <?php
+declare(strict_types=1);
+
 /**
  * @copyright Copyright (c) 2013-2015 2amigOS! Consulting Group LLC
- * @link http://2amigos.us
- * @license http://www.opensource.org/licenses/bsd-license.php New BSD License
+ * @link https://2amigos.us
+ * @license https://www.opensource.org/licenses/bsd-license.php New BSD License
  */
 
 namespace dosamigos\leaflet\types;
@@ -10,14 +12,13 @@ namespace dosamigos\leaflet\types;
 use yii\base\InvalidConfigException;
 use yii\web\JsExpression;
 
-
 /**
  * Bounds represents a rectangular area in pixel coordinates.
  *
- * @see http://leafletjs.com/reference.html#bounds
+ * @see https://leafletjs.com/reference.html#bounds
  * @author Antonio Ramirez <amigo.cobos@gmail.com>
- * @link http://www.ramirezcobos.com/
- * @link http://www.2amigos.us/
+ * @link https://www.ramirezcobos.com/
+ * @link https://www.2amigos.us/
  * @package dosamigos\leaflet\types
  */
 
@@ -31,17 +32,17 @@ class Bounds extends Type implements ArrayableInterface
     /**
      * @var Point the top left corner of the rectangle
      */
-    private $_min;
+    private ?Point $_min = null;
 
     /**
      * @var Point the bottom right corner of the rectangle
      */
-    private $_max;
+    private ?Point $_max = null;
 
     /**
      * @param Point $max
      */
-    public function setMax(Point $max)
+    public function setMax(Point $max): void
     {
         $this->_max = $max;
     }
@@ -49,7 +50,7 @@ class Bounds extends Type implements ArrayableInterface
     /**
      * @return Point
      */
-    public function getMax()
+    public function getMax(): ?Point
     {
         return $this->_max;
     }
@@ -57,7 +58,7 @@ class Bounds extends Type implements ArrayableInterface
     /**
      * @param Point $min
      */
-    public function setMin(Point $min)
+    public function setMin(Point $min): void
     {
         $this->_min = $min;
     }
@@ -65,7 +66,7 @@ class Bounds extends Type implements ArrayableInterface
     /**
      * @return Point
      */
-    public function getMin()
+    public function getMin(): ?Point
     {
         return $this->_min;
     }
@@ -74,7 +75,7 @@ class Bounds extends Type implements ArrayableInterface
      * Initializes the object
      * @throws \yii\base\InvalidConfigException
      */
-    public function init()
+    public function init(): void
     {
         if (empty($this->min) || empty($this->max)) {
             throw new InvalidConfigException("'min' and 'max' attributes cannot be empty.");
@@ -84,7 +85,7 @@ class Bounds extends Type implements ArrayableInterface
     /**
      * @return \yii\web\JsExpression the js initialization code of the object
      */
-    public function encode()
+    public function encode(): JsExpression
     {
         $min = $this->getMin()->toArray(true);
         $max = $this->getMax()->toArray(true);
@@ -99,7 +100,7 @@ class Bounds extends Type implements ArrayableInterface
      *
      * @return array the array representation of this object
      */
-    public function toArray($encode = false)
+    public function toArray($encode = false): mixed
     {
         $min = $this->getMin()->toArray($encode);
         $max = $this->getMax()->toArray($encode);
