@@ -1,8 +1,10 @@
 <?php
+declare(strict_types=1);
+
 /**
  * @copyright Copyright (c) 2013-2015 2amigOS! Consulting Group LLC
- * @link http://2amigos.us
- * @license http://www.opensource.org/licenses/bsd-license.php New BSD License
+ * @link https://2amigos.us
+ * @license https://www.opensource.org/licenses/bsd-license.php New BSD License
  */
 
 namespace dosamigos\leaflet\types;
@@ -15,10 +17,10 @@ use yii\web\JsExpression;
  *
  * DivIcon represents a lightweight icon for markers that uses a simple div element instead of an image.
  *
- * @see http://leafletjs.com/reference.html#divicon
+ * @see https://leafletjs.com/reference.html#divicon
  * @author Antonio Ramirez <amigo.cobos@gmail.com>
- * @link http://www.ramirezcobos.com/
- * @link http://www.2amigos.us/
+ * @link https://www.ramirezcobos.com/
+ * @link https://www.2amigos.us/
  * @package dosamigos\leaflet\types
  */
 class DivIcon extends Type
@@ -40,30 +42,34 @@ class DivIcon extends Type
      * L.marker({icon: L.icon({...}), ...).addTo(map);
      * ```
      */
-    public $name;
+    public ?string $name = null;
+
     /**
      * @var string a custom class name to assign to both icon and shadow images. Empty by default.
      */
-    public $className;
+    public ?string $className = null;
+
     /**
      * @var string a custom HTML code to put inside the div element, empty by default.
      */
-    public $html;
+    public ?string $html = null;
+
     /**
      * @var Point size of the icon image in pixels.
      */
-    private $_iconSize;
+    private ?Point $_iconSize = null;
+    
     /**
      * @var Point the coordinates of the "tip" of the icon (relative to its top left corner). The icon will be aligned so
      * that this point is at the marker's geographical location. Centered by default if size is specified, also can be
      * set in CSS with negative margins.
      */
-    private $_iconAnchor;
+    private ?Point $_iconAnchor = null;
 
     /**
      * @param Point $iconSize
      */
-    public function setIconSize(Point $iconSize)
+    public function setIconSize(Point $iconSize): void
     {
         $this->_iconSize = $iconSize;
     }
@@ -71,7 +77,7 @@ class DivIcon extends Type
     /**
      * @return Point
      */
-    public function getIconSize()
+    public function getIconSize(): ?Point
     {
         return $this->_iconSize;
     }
@@ -79,7 +85,7 @@ class DivIcon extends Type
     /**
      * @param Point $iconAnchor
      */
-    public function setIconAnchor(Point $iconAnchor)
+    public function setIconAnchor(Point $iconAnchor): void
     {
         $this->_iconAnchor = $iconAnchor;
     }
@@ -87,7 +93,7 @@ class DivIcon extends Type
     /**
      * @return Point
      */
-    public function getIconAnchor()
+    public function getIconAnchor(): ?Point
     {
         return $this->_iconAnchor;
     }
@@ -95,7 +101,7 @@ class DivIcon extends Type
     /**
      * @return \yii\web\JsExpression the js initialization code of the object
      */
-    public function encode()
+    public function encode(): JsExpression
     {
         $options = Json::encode($this->getOptions(), LeafLet::JSON_OPTIONS);
 
@@ -109,7 +115,7 @@ class DivIcon extends Type
     /**
      * @return array the configuration options of the array
      */
-    public function getOptions()
+    public function getOptions(): array
     {
         $options = [];
         $class = new \ReflectionClass(__CLASS__);
