@@ -1,8 +1,10 @@
 <?php
+declare(strict_types=1);
+
 /**
  * @copyright Copyright (c) 2013-2015 2amigOS! Consulting Group LLC
- * @link http://2amigos.us
- * @license http://www.opensource.org/licenses/bsd-license.php New BSD License
+ * @link https://2amigos.us
+ * @license https://www.opensource.org/licenses/bsd-license.php New BSD License
  */
 
 namespace dosamigos\leaflet\controls;
@@ -13,10 +15,10 @@ use yii\web\JsExpression;
  * Zoom renders basic zoom control with two buttons (zoom in and zoom out). It is put on the map by default unless you
  * set the map's zoomControl option to false
  *
- * @see http://leafletjs.com/reference.html#control-zoom
+ * @see https://leafletjs.com/reference.html#control-zoom
  * @author Antonio Ramirez <amigo.cobos@gmail.com>
- * @link http://www.ramirezcobos.com/
- * @link http://www.2amigos.us/
+ * @link https://www.ramirezcobos.com/
+ * @link https://www.2amigos.us/
  * @package dosamigos\leaflet\controls
  */
 class Zoom extends Control
@@ -25,7 +27,7 @@ class Zoom extends Control
      * Returns the javascript ready code for the object to render
      * @return \yii\web\JsExpression
      */
-    public function encode()
+    public function encode(): JsExpression
     {
         $this->clientOptions['position'] = $this->position;
         $options = $this->getOptions();
@@ -37,5 +39,4 @@ class Zoom extends Control
         }
         return new JsExpression($js);
     }
-
 }
